Fix compile error

+++ src/widgets/room/delegate/messageattachmentdelegatehelpertext.cpp
@@ -108,7 +108,7 @@ QSize MessageAttachmentDelegateHelperText::sizeHint(const MessageAttachment &msg
     if ((layout.isShown && !layout.title.isEmpty()) || layout.title.isEmpty()) {
         height += layout.textSize.height() + DelegatePaintUtil::margin();
     }
-    return {qMax(layout.titleSize.width(), (qreal)maxWidth),
+    return {static_cast<int>(qMax(layout.titleSize.width(), (qreal)maxWidth)),
                  height};
 }
